requirements.txt
django == 3.1.1
djangorestframework
django-cors-headers